Fixed S118 to add additional checks (BadAbstractClassName)
