--- src/main/java/lang/naming/badabstractclassnamecheck-normalclass.yaml
@@ -1,21 +1,0 @@
-rules:
-  - id: badabstractclassnamecheck-normalclass
-    pattern-either:
-      - patterns:
-          - pattern: class $ABSTRACTLIKE { ... }
-          - pattern-not: abstract class $ABSTRACTLIKE { ... }
-          - metavariable-regex:
-              metavariable: $ABSTRACTLIKE
-              regex: (^Abstract[A-Z][a-zA-Z0-9]*$)
-    message: Make this class abstract or rename it, since it matches the regular
-      expression (^Abstract[A-Z][a-zA-Z0-9]*$) without being abstract
-    languages:
-      - java
-    severity: WARNING
-    metadata:
-      key: S118
-      category: best-practice
-      references:
-        - https://github.com/SonarSource/sonar-java/blob/master/java-checks/src/main/java/org/sonar/java/checks/naming/BadAbstractClassNameCheck.java
-      technology:
-        - java
